+++ internal/controller/type/delete.ts
@@ -32,6 +32,10 @@ async function DeleteTypes(req: express.Request<any, any, any, IdsQuery>, res: e
     const prisma = new PrismaClient(PrismaClientOption)
 
     // query.ids 为字符串时转换为单元素数组, 为数组时无改变
+    if (query.ids === undefined) {
+        res.status(403).type("text/plain").send("ids not found in query")
+        return
+    }
     query.ids = [].concat(query.ids)
 
     await prisma.type.deleteMany({
